+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Threadjob
 - help menu
 - change colors
+- Add new frame with Azure resource details
 
 ### Changed